faiss/requirements.txt
faiss-cpu
fastapi
uvicorn
psycopg2-binary
numpy
python-dotenv
sqlalchemy
pandas
pydantic